# packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "CartDetail": {
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cardmarket Created At"
          },
          "cardmarket_id": {
            "title": "Cardmarket Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "raw_source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Source Reference"
          },
          "status": {
            "$ref": "#/components/schemas/CartStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cardmarket_id",
          "buyer_name",
          "status",
          "total_amount",
          "currency",
          "item_count",
          "cardmarket_created_at",
          "created_at",
          "updated_at",
          "last_seen_at",
          "items",
          "raw_source_reference"
        ],
        "title": "CartDetail",
        "type": "object"
      },
      "CartFilter": {
        "enum": [
          "all",
          "to_pay",
          "paid",
          "completed",
          "cancelled"
        ],
        "title": "CartFilter",
        "type": "string"
      },
      "CartListItem": {
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cardmarket Created At"
          },
          "cardmarket_id": {
            "title": "Cardmarket Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CartStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cardmarket_id",
          "buyer_name",
          "status",
          "total_amount",
          "currency",
          "item_count",
          "cardmarket_created_at",
          "created_at",
          "updated_at",
          "last_seen_at"
        ],
        "title": "CartListItem",
        "type": "object"
      },
      "CartStatus": {
        "enum": [
          "TO_PAY",
          "PAID",
          "COMPLETED",
          "CANCELLED",
          "UNKNOWN"
        ],
        "title": "CartStatus",
        "type": "string"
      },
      "CartsBatch": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/NormalizedCart"
            },
            "title": "Details",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/NormalizedCartSummary"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "CartsBatch",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "AUTH_ERROR",
          "NETWORK_ERROR",
          "CARDMARKET_CHANGED",
          "SELECTOR_NOT_FOUND",
          "TIMEOUT",
          "ACTION_FAILED",
          "VERIFICATION_FAILED",
          "DATABASE_ERROR",
          "UNKNOWN"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemOut": {
        "properties": {
          "card_name": {
            "title": "Card Name",
            "type": "string"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "expansion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expansion"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "total_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          }
        },
        "required": [
          "id",
          "card_name",
          "expansion",
          "language",
          "condition",
          "quantity",
          "unit_price",
          "total_price"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
      "NormalizedCart": {
        "additionalProperties": false,
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_id": {
            "minLength": 1,
            "title": "Cardmarket Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "currency": {
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NormalizedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/CartStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          }
        },
        "required": [
          "cardmarket_id",
          "buyer_name",
          "status"
        ],
        "title": "NormalizedCart",
        "type": "object"
      },
      "NormalizedCartSummary": {
        "additionalProperties": false,
        "properties": {
          "buyer_name": {
            "title": "Buyer Name",
            "type": "string"
          },
          "cardmarket_id": {
            "minLength": 1,
            "title": "Cardmarket Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "currency": {
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "item_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Count"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/CartStatus"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          }
        },
        "required": [
          "cardmarket_id",
          "buyer_name",
          "status"
        ],
        "title": "NormalizedCartSummary",
        "type": "object"
      },
...
      "NormalizedItem": {
        "additionalProperties": false,
        "properties": {
          "card_name": {
            "title": "Card Name",
            "type": "string"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "expansion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expansion"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "quantity": {
            "default": 1,
            "minimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "total_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          }
        },
        "required": [
          "card_name"
        ],
        "title": "NormalizedItem",
        "type": "object"
      },
...
      "Page_CartListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CartListItem]",
        "type": "object"
      },
...
      "Page_SyncRunOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SyncRunOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[SyncRunOut]",
        "type": "object"
      },
...
      "SyncBatchResult": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "records_changed": {
            "title": "Records Changed",
            "type": "integer"
          },
          "records_found": {
            "title": "Records Found",
            "type": "integer"
          }
        },
        "required": [
          "records_found",
          "records_changed",
          "events"
        ],
        "title": "SyncBatchResult",
        "type": "object"
      },
...
      "SyncEntity": {
        "enum": [
          "ORDERS",
          "CONVERSATIONS",
          "CARTS",
          "FULL"
        ],
        "title": "SyncEntity",
        "type": "string"
      },
      "SyncRunOut": {
        "properties": {
          "artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "entity_type": {
            "$ref": "#/components/schemas/SyncEntity"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "records_changed": {
            "title": "Records Changed",
            "type": "integer"
          },
          "records_found": {
            "title": "Records Found",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SyncRunStatus"
          },
          "sync_id": {
            "title": "Sync Id",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sync_id",
          "started_at",
          "completed_at",
          "status",
          "entity_type",
          "trigger",
          "records_found",
          "records_changed",
          "error",
          "error_code",
          "artifacts"
        ],
        "title": "SyncRunOut",
        "type": "object"
      },
      "SyncRunStatus": {
        "enum": [
          "RUNNING",
          "SUCCESS",
          "PARTIAL",
          "FAILED"
        ],
        "title": "SyncRunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cardmarket Companion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/carts": {
      "get": {
        "operationId": "list_carts_api_carts_get",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CartFilter",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CartListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Carts",
        "tags": [
          "carts"
        ]
      }
    },
    "/api/carts/{cart_id}": {
      "get": {
        "operationId": "get_cart_api_carts__cart_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cart",
        "tags": [
          "carts"
        ]
      }
    },
...
    "/api/sync/runs": {
      "get": {
        "operationId": "sync_runs_api_sync_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SyncRunOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Runs",
        "tags": [
          "connection"
        ]
      }
    },
...
    "/internal/agent/sync/{run_id}/carts": {
      "post": {
        "operationId": "sync_carts_internal_agent_sync__run_id__carts_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartsBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Carts",
        "tags": [
          "internal-agent"
        ]
      }
    },
...
  }
}